+++ backend/Split-backend/app.js
@@ -4,7 +4,6 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
-var indexRouter = require('./routes/index');
 var groupRouter = require('./routes/group');
 var loginRouter = require('./routes/login');
 var userRouter = require('./routes/user');
@@ -21,7 +20,6 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
 app.use('/group', groupRouter);
 app.use('/login', loginRouter);
 app.use('/user', userRouter);
